pyproject.toml
[project]
name = "datapilot"
version = "0.1.0"
description = "Enterprise Agentic BI Platform - Text-to-SQL with Multi-Agent Architecture"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "DataPilot Team" }
]

dependencies = [
    # 核心编排 (LangChain 1.x 最新版 - 2025.12)
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langchain-community>=0.4.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.0",

    # LLM (DeepSeek 使用 OpenAI 兼容接口)
    "openai>=1.40.0",
    "dspy-ai>=2.5.0",

    # Embedding (开源替代方案)
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",

    # 向量检索
    "lancedb>=0.15.0",
    "pyarrow>=18.0.0",
    "qdrant-client>=1.10.0",

    # SQL 处理
    "sqlglot>=25.0.0",
    "sqlalchemy>=2.0.0",

    # 核心数据库驱动 (MySQL/PostgreSQL/SQLite)
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.20.0",
    "pymysql>=1.1.0",

    # 沙箱 (E2B)
    "e2b>=1.0.0",
    "e2b-code-interpreter>=1.0.0",

    # 缓存
    "redis>=5.0.0",

    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",

    # 可观测性
    "prometheus-client>=0.20.0",
    "structlog>=24.0.0",

    # 工具
    "httpx>=0.27.0",
    "tenacity>=8.4.0",
    "python-dotenv>=1.0.0",
    "orjson>=3.10.0",
    "networkx>=3.3",
    "jieba>=0.42.0",
    "faker>=25.0.0",
    "langsmith>=0.1.20",
]

[project.optional-dependencies]
# 扩展数据库驱动 (按需安装)
sqlserver = [
    "aioodbc>=0.5.0",
    "pyodbc>=5.0.0",
]

clickhouse = [
    "clickhouse-connect>=0.8.0",
    "asynch>=0.2.0",
]

duckdb = [
    "duckdb>=1.0.0",
]

# 所有扩展数据库
all-db = [
    "aioodbc>=0.5.0",
    "pyodbc>=5.0.0",
    "clickhouse-connect>=0.8.0",
    "asynch>=0.2.0",
    "duckdb>=1.0.0",
]

dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "hypothesis>=6.100.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

eval = [
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "matplotlib>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/datapilot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/datapilot --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
